move deserialize_with_ext to server again
