+++ types/src/lib.rs
@@ -48,91 +48,3 @@ pub use http::Extensions;
 pub use params::{Id, InvalidRequestId, Params, ParamsSequence, SubscriptionId, TwoPointZero};
 pub use request::{InvalidRequest, Notification, Request};
 pub use response::{Response, ResponsePayload, SubscriptionPayload, SubscriptionResponse, Success as ResponseSuccess};
-
-/// Deserialize calls, notifications and responses with HTTP extensions.
-pub mod deserialize_with_ext {
-
-	/// Method call.
-	pub mod call {
-		use crate::Request;
-
-		/// Wrapper over `serde_json::from_slice` that sets the extensions.
-		pub fn from_slice<'a>(
-			data: &'a [u8],
-			extensions: &'a http::Extensions,
-		) -> Result<Request<'a>, serde_json::Error> {
-			let mut req: Request = serde_json::from_slice(data)?;
-			*req.extensions_mut() = extensions.clone();
-			Ok(req)
-		}
-
-		/// Wrapper over `serde_json::from_str` that sets the extensions.
-		pub fn from_str<'a>(data: &'a str, extensions: &'a http::Extensions) -> Result<Request<'a>, serde_json::Error> {
-			let mut req: Request = serde_json::from_str(data)?;
-			*req.extensions_mut() = extensions.clone();
-			Ok(req)
-		}
-	}
-
-	/// Notification.
-	pub mod notif {
-		use crate::Notification;
-
-		/// Wrapper over `serde_json::from_slice` that sets the extensions.
-		pub fn from_slice<'a, T>(
-			data: &'a [u8],
-			extensions: &'a http::Extensions,
-		) -> Result<Notification<'a, T>, serde_json::Error>
-		where
-			T: serde::Deserialize<'a>,
-		{
-			let mut notif: Notification<T> = serde_json::from_slice(data)?;
-			*notif.extensions_mut() = extensions.clone();
-			Ok(notif)
-		}
-
-		/// Wrapper over `serde_json::from_str` that sets the extensions.
-		pub fn from_str<'a, T>(
-			data: &'a str,
-			extensions: &http::Extensions,
-		) -> Result<Notification<'a, T>, serde_json::Error>
-		where
-			T: serde::Deserialize<'a>,
-		{
-			let mut notif: Notification<T> = serde_json::from_str(data)?;
-			*notif.extensions_mut() = extensions.clone();
-			Ok(notif)
-		}
-	}
-
-	/// Response.
-	pub mod response {
-		use crate::Response;
-
-		/// Wrapper over `serde_json::from_slice` that sets the extensions.
-		pub fn from_slice<'a, T>(
-			data: &'a [u8],
-			extensions: &'a http::Extensions,
-		) -> Result<Response<'a, T>, serde_json::Error>
-		where
-			T: serde::Deserialize<'a> + Clone,
-		{
-			let mut res: Response<T> = serde_json::from_slice(data)?;
-			*res.extensions_mut() = extensions.clone();
-			Ok(res)
-		}
-
-		/// Wrapper over `serde_json::from_str` that sets the extensions.
-		pub fn from_str<'a, T>(
-			data: &'a str,
-			extensions: &'a http::Extensions,
-		) -> Result<Response<'a, T>, serde_json::Error>
-		where
-			T: serde::Deserialize<'a> + Clone,
-		{
-			let mut res: Response<T> = serde_json::from_str(data)?;
-			*res.extensions_mut() = extensions.clone();
-			Ok(res)
-		}
-	}
-}
